requirements.txt
bcrypt==3.1.7
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
confluent-kafka==2.11.1
cryptography==2.8
future==0.18.2
httplib2==0.14.0
idna==2.8
paramiko==2.6.0
protobuf==3.6.1
psycopg2-binary==2.9.10
py4j==0.10.9.7
PyJWT==1.7.1
pyspark==3.5.1
python-dateutil==2.7.3
pytz==2019.3
PyYAML==5.3.1
requests==2.22.0
simplejson==3.16.0
six==1.14.0
ua-parser==0.18.0
user-agents==2.2.0
urllib3==1.25.8
venv-pack

